backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
selenium==4.11.2
webdriver-manager==4.0.0
python-dotenv==1.0.0
lxml==4.9.3
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1